New module kextensions with ContainerRecyclerViewHolder instead of using kotlin-android-extension's LayoutContainer in BaseRecyclerViewHolder Refactored sample app with new kextensions module Cleanup unused dependencies
